feat: custom chatId route support
