+++ src/lib/components/chat/MessageInput/Suggestions.svelte
